+++ src/main/java/sylenthuntress/aceofhearts/registry/ModAttachmentTypes.java
@@ -3,8 +3,6 @@ package sylenthuntress.aceofhearts.registry;
 import com.mojang.serialization.Codec;
 import net.fabricmc.fabric.api.attachment.v1.AttachmentRegistry;
 import net.fabricmc.fabric.api.attachment.v1.AttachmentType;
-import net.minecraft.util.math.BlockPos;
-import net.minecraft.util.math.Vec3d;
 import sylenthuntress.aceofhearts.AceOfHearts;
 
 @SuppressWarnings("UnstableApiUsage")
@@ -21,14 +19,6 @@ public class ModAttachmentTypes {
             builder -> builder
                     .initializer(() -> false)
                     .persistent(Codec.BOOL)
-    );
-
-    public static final AttachmentType<BlockPos> DEATH_COORDS = AttachmentRegistry.create(
-            AceOfHearts.modIdentifier("death_coords"),
-            builder -> builder
-                    .initializer(() -> BlockPos.ORIGIN)
-                    .persistent(BlockPos.CODEC)
-                    .copyOnDeath()
     );
 
     public static void registerAll() {